+++ src/global_functions.php
@@ -329,29 +329,19 @@ if (!function_exists('unlink_recursive')) {
      * @param string $dirname The directory path
      * @param array|bool $exceptions Either an array of files to exclude
      *  or boolean true to not grab dot files
-     * @return bool
+     * @return void
      * @since 1.0.7
      */
     function unlink_recursive($dirname, $exceptions = false)
     {
         list($directories, $files) = dir_tree($dirname, $exceptions);
 
-        //Adds links. `dir_tree()` returns links as directories
-        foreach ($directories as $directory) {
-            if (is_link($directory)) {
-                $files[] = $directory;
-            }
-        }
-
-        $success = true;
+        //Adds symlinks. `dir_tree()` returns symlinks as directories
+        $files += array_filter($directories, 'is_link');
 
         foreach ($files as $file) {
-            if (!unlink($file)) {
-                $success = false;
-            }
+            is_link($file) && is_dir($file) && is_win() ? rmdir($file) : unlink($file);
         }
-
-        return $success;
     }
 }